test: OMP-as-oracle differential ground truth for all stages (tree + cifar)
Replace the smoke-level gtest suite (no-throw / dims / "buffer changed";
0 numerical assertions, 0 cross-backend checks) with real per-stage
correctness checks. The CPU/OMP path is the oracle: a backend kernel is
correct iff its stage output matches OMP on the same seeded input.

Harness (backend-parametrized, reused by CUDA/Vulkan later via one macro):
- common/testing/oracle.hpp  ExactEqual (int/structural, first-mismatch) +
  NearEqual (float rtol/atol, worst-elem + max-diff, NaN/Inf-aware).
- common/testing/cnn_ref.hpp independent DOUBLE-precision conv2d/maxpool/
  linear reference (cifar dense + sparse share it).
- tree/tree_diff_oracle.hpp   7 stages exact/near; oracle pinned single-
  threaded (radix-tree/octree builds are order-nondeterministic under
  parallel omp-for); stage-7 children/leaf_mask excluded (multiple valid
  representations -> need invariant compare for cross-backend).
- cifar-{dense,sparse}_diff_oracle.hpp  9 stages each, per-stage isolated
  (ref fed the actual upstream buffer). cifar-sparse builds a deterministic
  valid CSR in-test because the shipped AppData leaves row_ptr/col_idx empty
  -> the shipped sparse pipeline computes all zeros (surfaced by a SKIP guard
  test ShippedCsrIndicesAreEmpty_KnownIssue).

Verified 25/25 stage checks on all four targets: x86 build box, Pixel 7a
(arm64), Jetson Orin (arm64), Samsung S24 (arm64) -- the differential oracle
holds across x86 and aarch64.

T0 fixes folded in:
- register octree test_sort.cpp + test_main.cpp (were xmake-only) in CMake;
  per-test CTest LABELS omp/cuda/vulkan for per-runner CI gating.
- parse_args_test(): non-fatal device parse so an unknown/missing --device
  self-skips instead of exit(1) aborting the binary (CI portability).
- fix cifar kTestBatchSize drift (-> AppData::BATCH_SIZE), tree uint32 morton
  read into vector<float>, and octree/appdata.hpp missing <algorithm>.

docs/TESTING.md: full (app x stage x backend x target) coverage plan.

+++ builtin-apps/tree/cuda/test_main.cu
@@ -15,12 +15,12 @@ TEST(Stage1Test, Basic) {
   tree::cuda::CudaDispatcher disp;
   tree::SafeAppData appdata(&disp.get_mr());
 
-  const std::vector<float> before(appdata.u_morton_keys_s1_out.begin(),
+  const std::vector<uint32_t> before(appdata.u_morton_keys_s1_out.begin(),
                                   appdata.u_morton_keys_s1_out.end());
 
   EXPECT_NO_THROW(disp.dispatch_stage(appdata, 1)) << "Stage 1 should not throw";
 
-  const std::vector<float> after(appdata.u_morton_keys_s1_out.begin(),
+  const std::vector<uint32_t> after(appdata.u_morton_keys_s1_out.begin(),
                                  appdata.u_morton_keys_s1_out.end());
 
   const bool is_different = !std::ranges::equal(before, after);
